+++ packages/util-crypto/src/base58/decode.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2020 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { assert, bufferToU8a } from '@polkadot/util';
+import { bufferToU8a } from '@polkadot/util';
 
 import { bs58 } from './bs58';
 import validate from './validate';
@@ -12,14 +12,8 @@ import validate from './validate';
  * @description
  * From the provided input, decode the base58 and return the result as an `Uint8Array`.
  */
-export default function base58Decode (value: string, ipfsCompat = false): Uint8Array {
-  if (ipfsCompat) {
-    assert(value[0] === 'z', "Expected IPFS base32 identifier 'z' at string start");
-
-    value = value.substr(1);
-  }
-
-  validate(value);
+export default function base58Decode (value: string, ipfsCompat?: boolean): Uint8Array {
+  validate(value, ipfsCompat);
 
   return bufferToU8a(bs58.decode(value));
 }