+++ packages/util-crypto/src/base58/validate.ts
@@ -1,6 +1,8 @@
 // Copyright 2017-2020 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { assert } from '@polkadot/util';
+
 import { BASE58_ALPHABET } from './bs58';
 
 export function validateChars (type: string, alphabet: string, value: string): true {
@@ -19,6 +21,12 @@ export function validateChars (type: string, alphabet: string, value: string): t
  * @description
  * Validates the the supplied value is valid base58
  */
-export default function base58Validate (value: string): true {
+export default function base58Validate (value: string, ipfsCompat = false): true {
+  if (ipfsCompat) {
+    assert(value[0] === 'z', "Expected IPFS base32 identifier 'z' at string start");
+
+    value = value.substr(1);
+  }
+
   return validateChars('base58', BASE58_ALPHABET, value);
 }